+++ packages/api/src/account-lcd.ts
@@ -1,0 +1,23 @@
+import LCDClient from './util/lcd'
+
+export type ICoin = {
+  denom: 'atto',
+  amount: string
+}
+
+export type IAccount = {
+  address: string
+  coins: ICoin[]
+  public_key: string
+  account_number: number
+  sequence: number
+}
+
+export default class Account extends LCDClient {
+
+  async get(address: string): Promise<IAccount> {
+    const account = (await this.query(`/auth/accounts/${address}`)).result.value
+    if (!account.address) account.address = address
+    return account
+  }
+}