+++ packages/api/src/util/resolve_test.ts
@@ -2,17 +2,19 @@ import { Test } from 'tape'
 import test from 'tape'
 import * as sinon from 'sinon'
 import { resolveSID } from './resolve'
-import Api from '../mock'
+import Api from '../lcd'
+import { IService } from '../service-lcd'
+import { IInstance } from '../instance-lcd'
 
-const instances = [{ hash: Buffer.from('instancehash'), serviceHash: Buffer.from('servicehash') }]
-const services = [{ hash: Buffer.from('servicehash'), sid: 'servicesid' }]
+const instances: IInstance[] = [{ hash: 'instancehash', serviceHash: 'servicehash', envHash: '' }]
+const services: IService[] = [{ hash: 'servicehash', sid: 'servicesid', configuration: {} }]
 
 test('resolve service invalid', async function (t: Test) {
   t.plan(1);
-  const api = Api('')
+  const api = new Api()
   const sid = "invalid"
   sinon.stub(api.instance, 'get').callsFake(() => { throw new Error("not found") })
-  sinon.stub(api.service, 'list').callsFake(() => (Promise.resolve({ services })))
+  sinon.stub(api.service, 'list').callsFake(() => Promise.resolve(services))
   try {
     await resolveSID(api, sid)
   } catch (e) {
@@ -22,21 +24,21 @@ test('resolve service invalid', async function (t: Test) {
 
 test('resolve service by sid', async function (t: Test) {
   t.plan(1);
-  const api = Api('')
+  const api = new Api()
   const sid = "servicesid"
   sinon.stub(api.instance, 'get').callsFake(() => { throw new Error("not found") })
-  sinon.stub(api.service, 'list').callsFake(() => (Promise.resolve({ services })))
-  sinon.stub(api.instance, 'list').callsFake(() => (Promise.resolve({ instances })))
+  sinon.stub(api.service, 'list').callsFake(() => Promise.resolve(services))
+  sinon.stub(api.instance, 'list').callsFake(() => Promise.resolve(instances))
   const instanceHash = await resolveSID(api, sid)
   t.equal(instanceHash, instances[0].hash)
 });
 
 test('resolve service by sid (multiple)', async function (t: Test) {
   t.plan(1);
-  const api = Api('')
+  const api = new Api()
   const sid = "multiplesid"
   sinon.stub(api.instance, 'get').callsFake(() => { throw new Error("not found") })
-  sinon.stub(api.service, 'list').callsFake(() => (Promise.resolve({ services: [{ sid }, { sid }] })))
+  sinon.stub(api.service, 'list').callsFake(() => Promise.resolve([{ sid, configuration: {} }, { sid, configuration: {} }]))
   try {
     await resolveSID(api, sid)
   } catch (e) {