+++ packages/cli/src/commands/service/list.ts
@@ -1,6 +1,5 @@
 import cli from 'cli-ux'
-import {IInstance} from '@mesg/api/lib/instance'
-import * as base58 from '@mesg/api/lib/util/base58'
+import {IInstance} from '@mesg/api/lib/instance-lcd'
 
 import Command from '../../root-command'
 
@@ -14,38 +13,38 @@ export default class ServiceList extends Command {
 
   async run(): Promise<IInstance[]> {
     const {flags} = this.parse(ServiceList)
-    const [{services}, {instances}, {runners}] = await Promise.all([
-      this.api.service.list({}),
-      this.api.instance.list({}),
-      this.api.runner.list({}),
+    const [services, instances, runners] = await Promise.all([
+      this.lcd.service.list(),
+      this.lcd.instance.list(),
+      this.lcd.runner.list(),
     ])
-    cli.table(services || [], {
-      hash: {header: 'HASH', get: srv => base58.encode(srv.hash)},
+    cli.table(services, {
+      hash: {header: 'HASH', get: srv => srv.hash},
       sid: {header: 'SID', get: srv => srv.sid},
       instances: {
         header: 'INSTANCES',
-        get: srv => (instances || [])
-          .filter(inst => inst.serviceHash.toString() === srv.hash.toString())
+        get: srv => instances
+          .filter(inst => inst.serviceHash === srv.hash)
           .map(inst => [
-            base58.encode(inst.hash),
-            (runners || [])
-              .filter(run => run.instanceHash.toString() === inst.hash.toString())
+            inst.hash,
+            runners
+              .filter(run => run.instanceHash === inst.hash)
               .reduce((p, _, i) => p + (i > 0 ? '\n' : ''), '')
           ].join(''))
         .join('\n'),
       },
       runners: {
         header: 'RUNNERS',
-        get: srv => (instances || [])
-          .filter(inst => inst.serviceHash.toString() === srv.hash.toString())
-          .map(inst => (runners || [])
-            .filter(run => run.instanceHash.toString() === inst.hash.toString())
-            .map(run => base58.encode(run.hash))
+        get: srv => instances
+          .filter(inst => inst.serviceHash === srv.hash)
+          .map(inst => runners
+            .filter(run => run.instanceHash === inst.hash)
+            .map(run => run.hash)
             .join('\n'),
           )
           .join('\n'),
       }
     }, {printLine: this.log, ...flags})
-    return instances || []
+    return instances
   }
 }