+++ packages/cli/src/commands/service/start.ts
@@ -1,9 +1,10 @@
-import {flags} from '@oclif/command'
-import {RunnerCreateOutputs} from '@mesg/api/lib/runner'
+import { flags } from '@oclif/command'
+import { RunnerCreateOutputs } from '@mesg/api/lib/runner'
 import * as base58 from '@mesg/api/lib/util/base58'
 
 import Command from '../../root-command'
-import {serviceResolver} from '../../utils/resolver'
+import { serviceResolver } from '../../utils/resolver'
+import { IRunner } from '@mesg/api/lib/runner-lcd'
 
 export default class ServiceStart extends Command {
   static description = 'Start a service by creating a new runner'
@@ -22,18 +23,19 @@ export default class ServiceStart extends Command {
     required: true,
   }]
 
-  async run(): RunnerCreateOutputs {
-    const {args, flags} = this.parse(ServiceStart)
+  async run(): Promise<IRunner> {
+    const { args, flags } = this.parse(ServiceStart)
     this.spinner.start('Starting runner')
     const serviceHash = await serviceResolver(this.api, args.SERVICE_HASH)
-    const {hash} = await this.api.runner.create({
+    const response = await this.api.runner.create({
       serviceHash,
       env: flags.env
     })
-    if (!hash) throw new Error('invalid runner')
-    this.spinner.stop(base58.encode(hash))
-    const runner = await this.api.runner.get({hash})
-    this.log(`Runner started with hash ${base58.encode(hash)} and instance hash ${base58.encode(runner.instanceHash)}`)
-    return {hash}
+    if (!response.hash) throw new Error('invalid runner')
+    const hash = base58.encode(response.hash)
+    this.spinner.stop(hash)
+    const runner = await this.lcd.runner.get(hash)
+    this.log(`Runner started with hash ${hash} and instance hash ${runner.instanceHash}`)
+    return runner
   }
 }